Refactored userprofile
